Update uint and define viewport using pixc namespace

+++ pixc/include/Foundation/Input/KeyCodes.h
@@ -6,7 +6,7 @@
  */
 namespace pixc {
 
-using KeyCode = unsigned int;
+using KeyCode = uint16_t;
 
 /**
  * @namespace pixc::key
